% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.gimlet.R
\name{clean.gimlet}
\alias{clean.gimlet}
\title{Preprocess Gimlet data.}
\usage{
clean.gimlet(data, tags, id)
}
\arguments{
\item{data}{A data frame with Gimlet-provided variable names.}

\item{tags}{Optional. A character vector of promoted tags.}

\item{id}{Optional. A character vector of users.}
}
\value{
A data frame containing processed data.
}
\description{
Preprocesses data output from Gimlet.
}
\examples{
clean.gimlet(read.gimlet("site", "e@mail.com", "pass"))

}
