% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimlet.prep.R
\name{gimlet.prep}
\alias{gimlet.prep}
\title{Prepare Gimlet data.}
\usage{
gimlet.prep(data, promoted_tags = NULL)
}
\arguments{
\item{data}{A data frame with raw Gimlet data.}

\item{promoted_tags}{Optional. A character vector of 'promoted' tags, i.e., tags to retain.}
}
\value{
A data frame with prepared data. If \code{promoted_tags} is provided, \code{tags} variable is replaced with \code{tagpre} (original tag) and \code{tagpost} (processed tag).
}
\description{
Prepare Gimlet data.
}
\examples{
prep.gimlet(read.gimlet('mysite', 'e@mail.com', 'mypassword'))
prep.gimlet(mydata, c('policy', 'technology', 'ill', 'internal_directions'))

}
