% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimlet.read.R
\name{gimlet.read}
\alias{gimlet.read}
\title{Download Gimlet data.}
\usage{
gimlet.read(site, email, password, start_date, end_date)
}
\arguments{
\item{site}{A string denoting the Gimlet subdomain 'site', e.g., YOURLIBRARY.gimlet.us.}

\item{email}{A string denoting the email to login to your Gimlet site.}

\item{password}{A string denoting the password to login to your Gimlet site.}

\item{start_date, end_date}{Optional. Object of class 'POSIXt' or 'Date' denoting the desired start and end date of the data query. Defaults to past 7 days.}
}
\value{
A data frame containing the contents of a Gimlet data query.
}
\description{
Download Gimlet data.
}
\examples{
read.gimlet('mysite', 'e@mail.com', 'mypassword')

}
