% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimlet.wc.R
\name{gimlet.wc}
\alias{gimlet.wc}
\title{Generate word clouds.}
\usage{
gimlet.wc(..., wc_opts = NULL)
}
\arguments{
\item{...}{Character vectors to include in the word cloud.}

\item{wc_opts}{A list of options to pass to \code{\link{wordcloud2}}.}
}
\value{
A png image of a wordcloud.
}
\description{
Generate word clouds.
}
\examples{
gimlet.wc(data$question, data$answer, wc_opts = list(color = 'red'))

}
