% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Welcome to gimletr!}
\usage{
hello(name)
}
\arguments{
\item{name}{Your name.}
}
\value{
A welcome message addressed to \code{name}.
}
\description{
Print a welcome message.
}
\examples{
hello("Tyler")
hello("Big T")

}
