% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.gimlet.R
\name{prep.gimlet}
\alias{prep.gimlet}
\title{Prepare Gimlet data.}
\usage{
prep.gimlet(data, promoted_tags = NULL)
}
\arguments{
\item{data}{A data frame with raw Gimlet data.}

\item{promoted_tags}{Optional. A character vector of 'promoted' tags, i.e., tags to retain.}
}
\value{
A data frame with prepared data.
}
\description{
Prepare Gimlet data.
}
\examples{
prep.gimlet(read.gimlet('mysite', 'e@mail.com', 'mypassword'))
prep.gimlet(mydata, c('policy', 'technology', 'ill', 'internal_directions'))

}
