% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.ids.R
\name{process.ids}
\alias{process.ids}
\title{Verify initials in Gimlet data.}
\usage{
process.ids(x, id)
}
\arguments{
\item{x, id}{Character vectors.}
}
\value{
\code{TRUE} if all elements in \code{data} are contained within \code{id}. \code{FALSE} and character vector of unmatched elements in there are any.
}
\description{
Verify initials in Gimlet data.
}
\examples{
process.ids(data$Initials, idAccepted)

}
