% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.initials.R
\name{process.initials}
\alias{process.initials}
\title{Verify initials in Gimlet data.}
\usage{
process.initials(initials, users)
}
\arguments{
\item{initials, users}{Character vectors.}
}
\value{
\code{TRUE} if all elements in \code{initials} are contained in \code{users}. \code{FALSE} otherwise.
}
\description{
Verify initials in Gimlet data.
}
\examples{
process.ids(initials = data$Initials, users = users)

}
