% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gimlet.R
\name{read.gimlet}
\alias{read.gimlet}
\title{Download Gimlet data.}
\usage{
read.gimlet(site, email, password, start_date, end_date)
}
\arguments{
\item{site, email, password}{Strings.}

\item{start_date, end_date}{Optional. Objects of class 'POSIXt' or 'Date'.}
}
\value{
A data frame containing the content of a Gimlet data query.
}
\description{
Download Gimlet data.
}
\examples{
read.gimlet('mysite', 'e@mail.com', 'mypassword')
read.gimlet('mysite', 'e@mail.com', 'mypassword', as.Date('2010-01-01'), as.Date('2010-01-31'))

}
