% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.errors.R
\name{compute.errors}
\alias{compute.errors}
\title{Compute user tagging errors.}
\usage{
compute.errors(initials, tagpre, tagpost, groups, ignore)
}
\arguments{
\item{initials, tagpre, tagpost}{Character vectors.}

\item{groups}{Optional. A character vector.}

\item{ignore}{Optional. An integer vector.}
}
\value{
A data frame containing \code{ent}, \code{err}, \code{per}, and \code{score} by each level in \code{initials}. If \code{groups} is specified, group errors are also computed. If \code{ignore} is specified, those indices are ignored when calculating \code{err}.
}
\description{
Compute user tagging errors.
}
\examples{
compute.errors(initials = data$initials, tagpre = data$tagpre, tagpost = data$tagpost, groups = data$group)

}
