% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimlet.errors.R
\name{gimlet.errors}
\alias{gimlet.errors}
\title{Compute user tagging errors.}
\usage{
gimlet.errors(initials, tagpre, tagpost)
}
\arguments{
\item{initials}{A character vector or factor denoting groups to compute errors for.}

\item{tagpre}{A character vector or factor denoting the submitted tags.}

\item{tagpost}{A character vector or factor denoting the corrected tags.}
}
\value{
A data frame containing \code{ent}, \code{err}, \code{per}, and \code{score}, grouped by \code{initials}.
}
\description{
Compute user tagging errors.
}
\examples{
compute.errors(data$initials, tagpre = data$tagpre, tagpost = data$tagpost)

}
