% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.gimlet.R
\name{prep.gimlet}
\alias{prep.gimlet}
\title{Pre-process Gimlet data.}
\usage{
prep.gimlet(data)
}
\arguments{
\item{data}{A data frame with raw Gimlet data.}
}
\value{
A data frame with pre-processed data.
}
\description{
Pre-process Gimlet data.
}
\examples{
prep.gimlet(data = read.gimlet('site', 'e@mail.com', 'pass'))

}
