% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.gimlet.R
\name{prep.gimlet}
\alias{prep.gimlet}
\title{Pre-process Gimlet data.}
\usage{
prep.gimlet(data)
}
\arguments{
\item{data}{A data frame (\code{\link[base]{data.frame}}) with Gimlet variables names.}
}
\value{
A data frame (\code{\link[base]{data.frame}}) containing pre-processed data.
}
\description{
Pre-processes Gimlet data.
}
\examples{
prep.gimlet(read.gimlet("site", "e@mail.com", "pass"))

}
