% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.ids.R
\name{process.ids}
\alias{process.ids}
\title{Verify Initials in Gimlet data.}
\usage{
process.ids(data, id)
}
\arguments{
\item{data}{A data frame (\code{\link[base]{data.frame}}) with Gimlet-provided variable names.}

\item{id}{A character vector of users.}
}
\value{
A data frame (\code{\link[base]{data.frame}}) containing processed and verified IDs.
}
\description{
Verify Initials in Gimlet data.
}
\examples{
process.ids(prep.gimlet(read.gimlet("mylibrary", "e@mail.com", "password")), ids)

}
