% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.tags.R
\name{process.tags}
\alias{process.tags}
\title{Verify tags in Gimlet data.}
\usage{
process.tags(tags, promoted, regex)
}
\arguments{
\item{tags, promoted, regex}{Character vectors.}
}
\value{
A vector of processed tags.
}
\description{
Verify tags in Gimlet data.
}
\examples{
process.tags(tags = data$Tags, promoted = c('policy', 'circulation'))

}
