% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.tags.R
\name{process.tags}
\alias{process.tags}
\title{Verify tags in Gimlet data.}
\usage{
process.tags(x, tags)
}
\arguments{
\item{x, tag}{Character vectors.}
}
\value{
A vector of processed tags.
}
\description{
Verify tags in Gimlet data.
}
\examples{
process.tags(data$Tags, tags)

}
