% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.tags.R
\name{process.tags}
\alias{process.tags}
\title{Verify tags in Gimlet data.}
\usage{
process.tags(tags, promoted, alt)
}
\arguments{
\item{tags, promoted, alt}{Character vectors.}
}
\value{
A vector of processed tags.
}
\description{
Verify tags in Gimlet data.
}
\examples{
process.tags(tags = data$Tags, promoted = c('policy', 'circulation'))
process.tags(tags = data$Tags, promoted = c('policy', 'circulation'), alt = list(policy = c('policies', 'hours'), circulation = c('circ', 'ill', 'borrowing')))

}
