% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.tags.R
\name{process.tags}
\alias{process.tags}
\title{Verify tags in Gimlet data.}
\usage{
process.tags(data, tags)
}
\arguments{
\item{data}{A data frame (\code{\link[base]{data.frame}}) with Gimlet-provided variable names.}

\item{tags}{A character vector of promoted tags.}
}
\value{
A data frame (\code{\link[base]{data.frame}}) containing processed and verified tags.
}
\description{
Verify tags in Gimlet data.
}
\examples{
process.tags(prep.gimlet(read.gimlet("mylibrary", "e@mail.com", "password")), tags)

}
