% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gimlet.R
\name{read.gimlet}
\alias{read.gimlet}
\title{Download Gimlet data.}
\usage{
read.gimlet(site, email, password, start_date, end_date)
}
\arguments{
\item{site}{The name of the Gimlet subdomain ("site") to access.}

\item{email}{The email address for logging in to the Gimlet site.}

\item{password}{The password for logging in to the Gimlet site.}

\item{start_date, end_date}{Optional. Objects of class "POSIXt" or "Date".}
}
\value{
A data frame (\code{\link[base]{data.frame}}) containing the data returned by Gimlet in the download query.
}
\description{
Download data from a Gimlet site.
}
\examples{
read.gimlet("mysite", "e@mail.com", "mypassword")
read.gimlet("mysite", "e@mail.com", "mypassword", "2015-01-01", "2015-01-31")
}
